+++ example/lib/pages/page2.dart
@@ -1,5 +1,4 @@
 import 'package:example/main.dart';
-
 import 'package:flutter/material.dart';
 import 'package:xly/xly.dart';